Rename main class tests
