+++ Audit-Service/src/test/java/io/github/arivanamin/scm/backend/audit/AuditApplicationIntegrationTest.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
-class AuditApplicationTest extends BaseDatabaseTest {
+class AuditApplicationIntegrationTest extends BaseDatabaseTest {
 
     @Test
     void mainShouldInitializeContext () {
