+++ Discovery-Server/src/test/java/io/github/arivanamin/scm/backend/discovery/DiscoveryServerApplicationIntegrationTest.java
@@ -1,12 +1,11 @@
 package io.github.arivanamin.scm.backend.discovery;
 
+import io.github.arivanamin.scm.backend.testing.architecture.bases.BaseIntegrationTest;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
 @Slf4j
-class DiscoveryServerApplicationTest {
+class DiscoveryServerApplicationIntegrationTest extends BaseIntegrationTest {
 
     @Test
     void mainShouldInitializeContext () {
