+++ Patient-Service/src/test/java/io/github/arivanamin/scm/backend/patient/PatientApplicationIntegrationTest.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
-class PatientApplicationTest extends BaseDatabaseTest {
+class PatientApplicationIntegrationTest extends BaseDatabaseTest {
 
     @Test
     void mainShouldInitializeContext () {
