+++ SSO-Service/src/test/java/io/github/arivanamin/scm/backend/sso/SsoApplicationIntegrationTest.java
@@ -1,11 +1,11 @@
-package io.github.arivanamin.scm.backend.apigateway;
+package io.github.arivanamin.scm.backend.sso;
 
 import io.github.arivanamin.scm.backend.testing.architecture.bases.BaseIntegrationTest;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 
 @Slf4j
-class ApiGatewayApplicationTest extends BaseIntegrationTest {
+class SsoApplicationIntegrationTest extends BaseIntegrationTest {
 
     @Test
     void mainShouldInitializeContext () {
